+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Moved `@ton-community/func-js` to `dependencies` to fix missing module error in downstream TypeScript projects.
+
 ## [0.35.1] - 2025-07-08
 
 ### Fixed