+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
 
 ### Added
 
 - Added coverage
+- Added `allowParallel` flag to `Blockchain.sendMessageIter` method, allowing for testing MITM attack
+- Added `autoDeployLibs` flag, that allow to detect libraries, deployed by contract in masterchain automatically
+
+### Changed
+
+- Added `@ton-community/func-js` to `peerDependencies` to fix missing module error in downstream TypeScript projects
 
 ## [0.35.1] - 2025-07-08