+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.37.1] - 2025-09-03
+
+### Fixed
+
+- Made `@ton/test-utils` import optional, preventing errors in environments where it's not installed.
+
 ## [0.37.0] - 2025-08-18
 
 ### Added